refactor: move to `utils`, use unconditional import
Resolves:
https://github.com/narwhals-dev/narwhals/pull/1955#discussion_r1945495203
https://github.com/narwhals-dev/narwhals/pull/1955#discussion_r1945496846
