+++ narwhals/utils.py
@@ -41,13 +41,19 @@ if TYPE_CHECKING:
     import pandas as pd
     from typing_extensions import Self
     from typing_extensions import TypeGuard
+    from typing_extensions import TypeIs
 
     from narwhals.dataframe import DataFrame
     from narwhals.dataframe import LazyFrame
     from narwhals.series import Series
+    from narwhals.typing import CompliantDataFrame
+    from narwhals.typing import CompliantLazyFrame
+    from narwhals.typing import CompliantSeries
+    from narwhals.typing import DataFrameLike
     from narwhals.typing import DTypes
     from narwhals.typing import IntoSeriesT
     from narwhals.typing import SizeUnit
+    from narwhals.typing import _NativeNamespace
 
     FrameOrSeriesT = TypeVar(
         "FrameOrSeriesT", bound=Union[LazyFrame[Any], DataFrame[Any], Series[Any]]
@@ -1170,3 +1176,23 @@ def check_column_names_are_unique(columns: list[str]) -> None:
         msg = "".join(f"\n- '{k}' {v} times" for k, v in duplicates.items())
         msg = f"Expected unique column names, got:{msg}"
         raise DuplicateError(msg)
+
+
+def is_compliant_dataframe(obj: Any) -> TypeIs[CompliantDataFrame]:
+    return hasattr(obj, "__narwhals_dataframe__")
+
+
+def is_compliant_lazyframe(obj: Any) -> TypeIs[CompliantLazyFrame]:
+    return hasattr(obj, "__narwhals_lazyframe__")
+
+
+def is_compliant_series(obj: Any) -> TypeIs[CompliantSeries]:
+    return hasattr(obj, "__narwhals_series__")
+
+
+def has_native_namespace(obj: Any) -> TypeIs[_NativeNamespace]:
+    return hasattr(obj, "__native_namespace__")
+
+
+def is_dataframe_like(obj: Any) -> TypeIs[DataFrameLike]:
+    return hasattr(obj, "__dataframe__")
